Fix variable was not parsed into VariableSegment
